Switch from mdns to nbonjour because node v10
