+++ lib/httpServer.js
@@ -6,8 +6,6 @@ const cors = require('@koa/cors')
 const route = require('koa-route')
 const static_ = require('koa-static')
 
-const mdns = require('mdns')
-
 const opts = require('minimist')(process.argv.slice(2))
 
 const log = () => async (context, next) => {
@@ -72,10 +70,9 @@ app.use(static_('examples'))
 
 module.exports = client => {
   app.context.client = client
-  const serviceType = mdns.makeServiceType(['http', 'tcp', 'livingroom'])
-
-  mdns.createAdvertisement(serviceType, 3000)
-  app.context.ad = mdns.createAdvertisement(serviceType, 3000)
-  app.context.ad.start()
+  // should be app.on('listening', () => {
+  const name = process.env.LIVING_ROOM_NAME || require('os').hostname()
+  const bonjour = require('nbonjour').create()
+  bonjour.publish({name, type: 'http', protocol: 'tcp', subtypes: ['livingroom'], port: 3000})
   return app
 }
