+++ lib/oscServer.js
@@ -45,9 +45,9 @@ class LivingRoomOscServer {
     const osc = new UDPPort({ localAddress, localPort })
 
     osc.on(`ready`, () => {
-      const mdns = require('mdns')
-      this.ad = mdns.createAdvertisement(mdns.udp('osc', 'livingroom'), localPort)
-      this.ad.start()
+      const name = process.env.LIVING_ROOM_NAME || require('os').hostname()
+      const bonjour = require('nbonjour').create()
+      bonjour.publish({name, type: 'osc', protocol: 'udp', subtypes: ['livingroom'], port: localPort})
     })
 
     osc.on('message', ({address, args}, _, {address: remoteAddress}) => {
