+++ lib/socketServer.js
@@ -1,6 +1,5 @@
 const Socket = require('koa-socket')
 const io = new Socket({})
-const mdns = require('mdns')
 
 io.on('assert', async ({ client, data: facts, acknowledge }) => {
   if (!Array.isArray(facts)) facts = [facts]
@@ -40,8 +39,6 @@ module.exports = (app, { verbose }) => client => {
   }
 
   const subscriptions = new Set()
-  const serviceType = mdns.makeServiceType('socketio', 'tcp', 'livingroom')
-  const ad = mdns.createAdvertisement(serviceType, 3000)
 
   io.use(async (context, next) => {
     context.client = client
@@ -50,7 +47,12 @@ module.exports = (app, { verbose }) => client => {
     await next()
   })
 
-  ad.start() // should really be on listen()
   io.attach(app)
+
+  // FIXME: wrap app.listen to do this
+  const name = process.env.LIVING_ROOM_NAME || require('os').hostname()
+  const bonjour = require('nbonjour').create()
+  bonjour.publish({name, type: 'socketio', protocol: 'tcp', subtypes: ['livingroom'], port: 3000})
+
   return app
 }
